refactor: Modularize webauthn functionality
Ideas for the new functionality hints that it makes more sense to
separate the core Keystone functionality and the extensions from each
other. For the webauthn it means that the core keystone state does not
need to keep the `Webauthn` object. Such split helps to keep things more
modular, inject them into the API at any point (more dynamically), make
them a features and even separating them into separate crates.
