+++ src/api/error.rs
@@ -129,12 +129,6 @@ pub enum KeystoneApiError {
     TokenError { source: TokenProviderError },
 
     #[error(transparent)]
-    WebAuthN {
-        #[from]
-        source: WebauthnError,
-    },
-
-    #[error(transparent)]
     Uuid {
         #[from]
         source: uuid::Error,
@@ -322,36 +316,6 @@ pub enum TokenError {
         #[from]
         source: crate::identity::types::user::DomainBuilderError,
     },
-}
-
-#[derive(Error, Debug)]
-pub enum WebauthnError {
-    #[error("unknown webauthn error")]
-    Unknown,
-    #[error("corrupt session")]
-    CorruptSession,
-    #[error("no session")]
-    NoSession,
-    #[error("User Not Found")]
-    UserNotFound,
-    #[error("User Has No Credentials")]
-    UserHasNoCredentials,
-}
-
-impl IntoResponse for WebauthnError {
-    fn into_response(self) -> Response {
-        let body = match self {
-            WebauthnError::CorruptSession => "Corrupt session",
-            WebauthnError::NoSession => "Registration state for the user not found",
-            WebauthnError::UserNotFound => "User Not Found",
-            WebauthnError::Unknown => "Unknown Error",
-            WebauthnError::UserHasNoCredentials => "User Has No Credentials",
-        };
-
-        // its often easiest to implement `IntoResponse` by calling other
-        // implementations
-        (StatusCode::INTERNAL_SERVER_ERROR, body).into_response()
-    }
 }
 
 impl From<AuthenticationError> for KeystoneApiError {
