+++ src/api/v4/token/restriction/delete.rs
@@ -43,7 +43,7 @@ use crate::token::TokenApi;
         (status = 404, description = "Token restriction not found", example = json!(KeystoneApiError::NotFound(String::from("id = 1"))))
     ),
     security(("x-auth" = [])),
-    tag="token_restrictions"
+    tag="token_restriction"
 )]
 #[tracing::instrument(
     name = "api::token_restriction::delete",
