+++ src/lib.rs
@@ -91,6 +91,7 @@ pub mod provider;
 pub mod resource;
 pub mod revoke;
 pub mod token;
+pub mod webauthn;
 
 #[cfg(test)]
 mod tests;
